updating project component
